omg-social-app: some refactor and extract or make new component from messagebox component
